refactor : eunbin 이 말한거 고친거
